+++ pkgs/test_descriptor/lib/src/utils.dart
@@ -9,6 +9,7 @@ import 'package:path/path.dart' as p;
 import 'package:term_glyph/term_glyph.dart' as glyph;
 import 'package:test/test.dart';
 
+import 'descriptor.dart';
 import 'sandbox.dart';
 
 /// A UTF-8 codec that allows malformed byte sequences.
@@ -24,6 +25,27 @@ String addBullet(String text) =>
 
 /// Converts [strings] to a bulleted list.
 String bullet(Iterable<String> strings) => strings.map(addBullet).join("\n");
+
+/// Returns a human-readable description of a directory with the given [name]
+/// and [contents].
+String describeDirectory(String name, List<Descriptor> contents) {
+  if (contents.isEmpty) return name;
+
+  var buffer = StringBuffer();
+  buffer.writeln(name);
+  for (var entry in contents.take(contents.length - 1)) {
+    var entryString = prefixLines(entry.describe(), '${glyph.verticalLine}   ',
+        first: '${glyph.teeRight}${glyph.horizontalLine}'
+            '${glyph.horizontalLine} ');
+    buffer.writeln(entryString);
+  }
+
+  var lastEntryString = prefixLines(contents.last.describe(), '    ',
+      first: '${glyph.bottomLeftCorner}${glyph.horizontalLine}'
+          '${glyph.horizontalLine} ');
+  buffer.write(lastEntryString);
+  return buffer.toString();
+}
 
 /// Prepends each line in [text] with [prefix].
 ///
@@ -67,7 +89,7 @@ bool matchesAll(Pattern pattern, String string) =>
 
 /// Like [Future.wait] with `eagerError: true`, but reports errors after the
 /// first using [registerException] rather than silently ignoring them.
-Future waitAndReportErrors(Iterable<Future> futures) {
+Future<List<T>> waitAndReportErrors<T>(Iterable<Future<T>> futures) {
   var errored = false;
   return Future.wait(futures.map((future) {
     // Avoid async/await so that we synchronously add error handlers for the